+++ qinclude/CQSVGFeTurbulence.h
@@ -13,7 +13,7 @@ class CQSVGFeTurbulence : public CQSVGObject, public CSVGFeTurbulence {
   Q_PROPERTY(double baseFreqX    READ getBaseFreqX   WRITE setBaseFreqX   )
   Q_PROPERTY(double baseFreqY    READ getBaseFreqY   WRITE setBaseFreqY   )
   Q_PROPERTY(int    numOctaves   READ getNumOctaves  WRITE setNumOctaves  )
-  Q_PROPERTY(int    seed         READ getSeed        WRITE setSeed        )
+  Q_PROPERTY(double seed         READ getSeed        WRITE setSeed        )
 
  public:
   CQSVGFeTurbulence(CQSVG *svg);
